requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
psycopg2-binary
python-dotenv
PyJWT
pwdlib
email-validator
python-multipart
